requirements.txt
pandas==1.1.3
requests==2.24.0
numpy==1.19.1
requests_html==0.10.0
IMDbPY==2020.9.25
scikit_learn==0.23.2
